+++ src/commands/Music/6_clear.js
@@ -1,6 +1,8 @@
 const Commando = require("discord.js-commando");
 const musicPlayerInstance = require("../../utils/music/musicPlayerInstance");
 const Discordcollection = require("../../utils/misc/Discordcollection");
+const { ErrorEmbed, SuccessEmbed } = require("../../utils/embed");
+const checkUserVc = require("../../utils/checkUserVc");
 module.exports = class AddCommand extends Commando.Command {
   constructor(client) {
     super(client, {
@@ -12,12 +14,31 @@ module.exports = class AddCommand extends Commando.Command {
     });
   }
   async run(message) {
+    const { channel } = message;
+    const id = message.member.user.id;
+
+    let voiceConnection = message.guild.me.voice.connection;
+
+    if (!voiceConnection) {
+      return ErrorEmbed("Queue is already empty.", channel);
+    }
+
+    const voiceChannelMembers = message.guild.me.voice.channel.members;
+
+    const isUserInVC = checkUserVc(voiceChannelMembers, id);
+
+    if (!isUserInVC)
+      return ErrorEmbed(
+        `<@${id}> You must be in the same voice channel to use this command.`,
+        channel
+      );
     const musicPlayer = musicPlayerInstance(message.channel);
     if (musicPlayer.isQueueEmpty() === true) {
-      return message.reply("Queue is already empty.");
+      return ErrorEmbed("Queue is already empty.", channel);
     }
 
     Discordcollection.delete(`MusicPlayer-${message.channel.guild.id}`);
-    message.reply("Queue cleared.");
+
+    SuccessEmbed("Queue cleared.", channel);
   }
 };
